# [timeseries] Speed up the train/val splitter