--- a/timeseries/src/autogluon/timeseries/dataset/ts_dataframe.py
+++ b/timeseries/src/autogluon/timeseries/dataset/ts_dataframe.py
@@ -517,15 +517,9 @@ class TimeSeriesDataFrame(pd.DataFrame):
             start_index = time_step_slice.start
             end_index = time_step_slice.stop
 
-        num_timesteps_per_item = self.num_timesteps_per_item()
-        # Create a boolean index that selects the correct slice in each timeseries
-        boolean_indicators = []
-        for length in num_timesteps_per_item:
-            indicator = np.zeros(length, dtype=bool)
-            indicator[start_index:end_index] = True
-            boolean_indicators.append(indicator)
-        index = np.concatenate(boolean_indicators)
-        result = TimeSeriesDataFrame(self[index].copy(), static_features=self.static_features)
+        time_step_slice = slice(start_index, end_index)
+        result = self.groupby(level=ITEMID, sort=False, as_index=False).nth(time_step_slice)
+        result.static_features = self.static_features
         result._cached_freq = self._cached_freq
         return result