*Description of changes:*
- Speed up the `append_suffix_to_item_id` function
- Replace the surprisingly slow `DataFrame.loc[index]` operation with `DataFrame.query("item_id in @index")`.

Testing on a subset of 5000 items from the M5 competition dataset:

Using code currently on `master`:
```
Loaded dataset with 7559974 rows and 5000 items.
df.slice_by_timestep(None, -prediction_length): 5.75s
LastWindowSplitter.split(df, prediction_length): 63.83s
```

After current PR:
```
Loaded dataset with 7559974 rows and 5000 items.
df.slice_by_timestep(None, -prediction_length): 1.38s
LastWindowSplitter.split(df, prediction_length): 8.30s
```

<details>
  <summary>Code for reproducing the results</summary>

```python
import time
import pandas as pd
from autogluon.timeseries import TimeSeriesDataFrame
from autogluon.timeseries.splitter import LastWindowSplitter

prediction_length = 28
# Dataset consists of the first 5000 items of the M5 competition dataset
raw_data = pd.read_parquet("../m5/data/subset.parquet")
static = pd.read_parquet("../m5/data/static.parquet")

raw_data["item_id"] = raw_data["item_id"].astype("str")
static["item_id"] = static["item_id"].astype("str")
static.set_index("item_id", inplace=True)

print(f"Loaded dataset with {len(raw_data)} rows and {raw_data['item_id'].nunique()} items.")
df = TimeSeriesDataFrame(raw_data, static_features=static)

start = time.time()
df.slice_by_timestep(None, -prediction_length)
print(f"df.slice_by_timestep(None, -prediction_length): {time.time() - start:.2f}s")

start = time.time()
splitter = LastWindowSplitter()
train_data, val_data = splitter.split(df, prediction_length)
print(f"LastWindowSplitter.split(df, prediction_length): {time.time() - start:.2f}s")
```

</details>



By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.